+++ AtwaterMonitor.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{F1569374-E546-4C2F-9428-3115EF32798A}</ProjectGuid>
     <OutputType>Exe</OutputType>
-    <RootNamespace>sandbox2</RootNamespace>
-    <AssemblyName>sandbox2</AssemblyName>
+    <RootNamespace>AtwaterMonitor</RootNamespace>
+    <AssemblyName>AtwaterMonitor</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -43,8 +43,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Program.cs" />
+    <Compile Include="Controller.cs" />
+    <Compile Include="Model.cs" />
+    <Compile Include="NetworkDevice.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="UPS.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
