+++ AtwaterMonitor.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29009.5
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sandbox2", "sandbox2.csproj", "{F1569374-E546-4C2F-9428-3115EF32798A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AtwaterMonitor", "AtwaterMonitor.csproj", "{F1569374-E546-4C2F-9428-3115EF32798A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
